osversion: Make OSVersion cross-platform
Move non-Windows dependent code outside of the `windows` build tagged
files. This makes it possible to use the version parse/stringifying code
on non-Windows platforms and keep the code that interacts with Windows
API only for the Windows builds.
