+++ osversion/osversion_windows.go
@@ -8,15 +8,6 @@ import (
 	"golang.org/x/sys/windows/registry"
 )
 
-// OSVersion is a wrapper for Windows version information
-// https://msdn.microsoft.com/en-us/library/windows/desktop/ms724439(v=vs.85).aspx
-type OSVersion struct {
-	Version      uint32
-	MajorVersion uint8
-	MinorVersion uint8
-	Build        uint16
-}
-
 var (
 	osv  OSVersion
 	once sync.Once
@@ -43,12 +34,6 @@ func Get() OSVersion {
 // The calling application must be manifested to get the correct version information.
 func Build() uint16 {
 	return Get().Build
-}
-
-// String returns the OSVersion formatted as a string. It implements the
-// [fmt.Stringer] interface.
-func (osv OSVersion) String() string {
-	return fmt.Sprintf("%d.%d.%d", osv.MajorVersion, osv.MinorVersion, osv.Build)
 }
 
 // ToString returns the OSVersion formatted as a string.
